backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.9
pydantic==2.10.6
numpy==2.2.1
Pillow==11.0.0
opencv-python-headless==4.10.0.84
Flask==3.1.0
streamlit==1.41.1
requests==2.32.3
pymongo==4.10.1
python-dotenv==1.0.1

# ML libraries to be integrated later (commented for now to keep installs lighter)
# torch
# torchvision
tensorflow
# keras

